+++ include/minizinc/hash.hh
@@ -352,6 +352,7 @@ public:
   /// Remove all elements from the map
   void clear() { _s.clear(); }
   bool isEmpty() const { return _s.begin() == _s.end(); }
+  unsigned int size() const { return _s.size(); }
 };
 
 }  // namespace MiniZinc